+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.4.6 (Next)
 
 * Extracted `SlackRubyBot::Server` from `SlackRubyBot::App` - [@dblock](https://github.com/dblock).
+* Fix: explicitly require 'giphy' - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.4.5 (10/29/2015)